Factor out server into a Cargo feature
