+++ src/constants.rs
@@ -1,0 +1,7 @@
+pub const STATIC_PUBLIC_SEED: [u8; 32] = [0u8; 32];
+pub const SEED_0: u8 = 0;
+pub const SEED_1: u8 = 1;
+
+pub const STATIC_SEED_2: [u8; 32] = [
+    2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+];
